Clean up function names & signatures

+++ src/common/const.jl
@@ -1,1 +1,3 @@
-const Bohr = 0.52917721092
+# Bohr radius in Angstrom unit
+# This is the default CODATA2006 value in W90 src/constants.F90
+const Bohr::Float64 = 0.52917721092
